requirements/base.txt
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
bcrypt~=4.0.1
boto3~=1.28.43
orjson~=3.9.1
psycopg[c]==3.1.12  # https://github.com/psycopg/psycopg
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.5  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.24.0.post1  # https://github.com/encode/uvicorn

# Django
# --------------------------
django~=4.2.3
django-allauth==0.58.2  # https://github.com/pennersr/django-allauth
django-cors-headers==4.3.0  # https://github.com/adamchainz/django-cors-headers

django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-filter~=23.2  # https://github.com/carltongibson/django-filter/
django-model-utils~=4.3.1  # https://github.com/jazzband/django-model-utils
django-redis==5.4.0  # https://github.com/jazzband/django-redis

djangorestframework==3.14.0
# DRF-spectacular for api documentation
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular
drf-orjson-renderer~=1.7.1
